refactor(material): extract DefineForm.svelte from MaterialPopup
Move the define-new-material form — toggle, composition input, parse
preview, name/density fields, save/use buttons — into its own component.
Parse helpers (parseMaterialInput, parseMassRatio) and associated
derivations/state move with it.

DefineForm owns its own defineOpen state plus the form fields. Parent
communicates edit intent via a reactive editInitial prop: set to an
EditableMaterial to open+seed, set to null to reset+collapse.
oncommit(material, enrichment) fires when the user commits (save & use
or use-without-saving); parent closes the modal in response.

Shell now ~100 LOC — pure wiring. No behavior change.

Refs: #64 Phase 0
